compilation failures
91>D:\winx64-debug\build\src\sql\sql_admin.cc(1754,43): warning C4267: 'argument': conversion from 'size_t' to 'uint', possible loss of data [D:\winx64-debug\build\sql\sql.vcxproj]
        key_unpack(&rec, table, this_key, fk_parts);
                                          ^

/home/buildbot/buildbot/build/mariadb-11.8.0/sql/sql_admin.cc:1734:35: error: comparison of integer expressions of different signedness: ‘ptrdiff_t’ {aka ‘int’} and ‘uint’ {aka ‘unsigned int’}
 1734 |   DBUG_ASSERT(keynr >= 0 && keynr < ref_table->s->keys);
      |                             ~~~~~~^~~~~~~~~~~~~~~~~~~~

+++ sql/key.h
@@ -35,7 +35,7 @@ void key_restore(uchar *to_record, const uchar *from_key, KEY *key_info,
 bool key_cmp_if_same(TABLE *form,const uchar *key,uint index,uint key_length);
 void key_unpack(String *to, TABLE *table, KEY *key);
 void key_unpack(String *to, const TABLE *table, const KEY *key,
-                uint prefix_size);
+                size_t prefix_size);
 void field_unpack(String *to, Field *field, const uchar *rec, uint max_length,
                   bool prefix_key);
 bool is_key_used(TABLE *table, uint idx, const MY_BITMAP *fields);
